Updated readme
- readme updated to show details of every entry in the BAB.
- code cleanup, removing old comments.
